+++ packages/rlc-common/src/metadata/buildPackageFile.ts
@@ -15,6 +15,7 @@ import { buildAzureMonorepoPackage } from "./packageJson/buildAzureMonorepoPacka
 import { buildAzureStandalonePackage } from "./packageJson/buildAzureStandalonePackage.js";
 import { buildFlavorlessPackage } from "./packageJson/buildFlavorlessPackage.js";
 import { getRelativePartFromSrcPath } from "../helpers/pathUtils.js";
+import { getPackageName } from "./utils.js";
 
 interface PackageFileOptions {
   exports?: Record<string, any>;
@@ -37,9 +38,7 @@ export function buildPackageFile(
     exports,
     azureArm: model.options?.azureArm,
     isModularLibrary: model.options?.isModularLibrary ?? false,
-    azureSdkForJs: model.options?.azureSdkForJs,
-    //TODO should remove this after finish the release tool test
-    shouldUsePnpmDep: model.options?.shouldUsePnpmDep
+    azureSdkForJs: model.options?.azureSdkForJs
   };
 
   let packageInfo: Record<string, any> = buildFlavorlessPackage(config);
@@ -111,15 +110,8 @@ export function updatePackageFile(
   if (hasLro) {
     packageInfo.dependencies = {
       ...packageInfo.dependencies,
-      // TODO remove model.options?.shouldUsePnpmDep after pnpm migration
-      "@azure/core-lro":
-        model.options?.shouldUsePnpmDep && model.options.azureSdkForJs
-          ? "workspace:^"
-          : "^3.1.0",
-      "@azure/abort-controller":
-        model.options?.shouldUsePnpmDep && model.options.azureSdkForJs
-          ? "workspace:^"
-          : "^2.1.2"
+      "@azure/core-lro": "^3.1.0",
+      "@azure/abort-controller": "^2.1.2"
     };
   }
 
@@ -139,10 +131,6 @@ function getDescription(model: RLCModel): string {
     return `A generated SDK for ${model.libraryName}.`;
   }
   return description;
-}
-
-function getPackageName(model: RLCModel): string {
-  return model.options?.packageDetails?.name ?? model.libraryName;
 }
 
 function getClientFilePath(model: RLCModel) {